+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Changed
 
 - Logged admin can see any project (#262)
 - Simpler and safer authentication logic (#270)
+- Use token based auth to reset passwords (#269)
 - Better install doc (#275)
 
 Added